requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
bleach==3.3.0
CacheControl==0.12.6
cachetools==4.2.2
celery==5.0.5
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-imagekit==4.0.2
django-import-export==2.5.0
django-js-asset==1.2.2
django-localflavor==3.0.1
django-mptt==0.12.0
django-redis==4.12.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.6.0
drf-api-tracking==1.8.0
drf-spectacular==0.16.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
Faker==8.2.0
fcm-django==1.0.1
firebase-admin==5.0.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
jsonschema==3.2.0
kombu==5.0.2
lml==0.1.0
MarkupPy==1.14
MarkupSafe==2.0.0
msgpack==1.0.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pilkit==2.0
Pillow==8.2.0
prompt-toolkit==3.0.18
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-stdnum==1.16
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
tablib==3.0.0
text-unidecode==1.3
twilio==6.61.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
pylibmc==1.6.1
sheet2dict==0.1.1